[MIG] account_payment_pro_multi_store: Migration to 17.0
