--- account_payment_group_multi_store/models/__init__.py
@@ -1,4 +1,0 @@
-from . import account_payment_group
-from . import account_payment
-from . import res_store
-from . import account_move
