--- account_payment_group_multi_store/views/account_payment_group_views.xml
@@ -1,30 +1,0 @@
-<?xml version="1.0"?>
-<odoo>
-    <!-- Customer Voucher Search -->
-    <record id="view_account_payment_group_form" model="ir.ui.view">
-        <field name="name">account.payment.group.form</field>
-        <field name="model">account.payment.group</field>
-        <field name="inherit_id" ref="account_payment_group.view_account_payment_group_form"/>
-        <field name="arch" type="xml">
-            <field name="company_id" position="after">
-                <!-- lo hacemos requerido por vista porque es computado (se computa luego de crear) y ademas por si al instalar algun store no se puede definir, ademas lo hacemos solo requerido en estado borrador ya que si queda algun pago validado sin este campo, dejamos que se reabra y de ultima se seleccione -->
-                <field name="store_id" groups="base_multi_store.group_multi_store" attrs="{'required': [('state', '=', 'draft')], 'readonly': ['|', ('state', '!=', 'draft'), ('payment_ids', '!=', [])]}" options="{'no_create': True, 'no_open': True}" invisible="context.get('pop_up')" force_save="True"/>
-            </field>
-        </field>
-    </record>
-
-    <record id="view_account_payment_group_search" model="ir.ui.view">
-        <field name="name">account.payment.group.select</field>
-        <field name="model">account.payment.group</field>
-        <field name="inherit_id" ref="account_payment_group.view_account_payment_group_search"/>
-        <field name="arch" type="xml">
-            <field name="company_id" position="after">
-                <field name="store_id" groups="base_multi_store.group_multi_store"/>
-            </field>
-            <search>
-                <filter name="store" string="Store" context="{'group_by':'store_id'}" groups="base_multi_store.group_multi_store"/>
-            </search>
-        </field>
-    </record>
-
-</odoo>
