+++ account_payment_pro_multi_store/models/__init__.py
@@ -1,0 +1,2 @@
+from . import account_payment
+from . import account_move
