+++ account_payment_pro_multi_store/views/account_payment_views.xml
@@ -1,0 +1,15 @@
+<?xml version="1.0"?>
+<odoo>
+    <!-- Customer Voucher Search -->
+    <record id="view_account_payment_form" model="ir.ui.view">
+        <field name="name">account.payment.form</field>
+        <field name="model">account.payment</field>
+        <field name="inherit_id" ref="account.view_account_payment_form"/>
+        <field name="arch" type="xml">
+            <field name="partner_bank_id" position="after">
+                <!-- lo hacemos requerido por vista porque es computado (se computa luego de crear) y ademas por si al instalar algun store no se puede definir, ademas lo hacemos solo requerido en estado borrador ya que si queda algun pago validado sin este campo, dejamos que se reabra y de ultima se seleccione -->
+                <field name="store_id" groups="base_multi_store.group_multi_store" required="state == 'draft'" readonly ="state != 'draft'" options="{'no_create': True, 'no_open': True}" invisible="context.get('pop_up')" force_save="True"/>
+            </field>
+        </field>
+    </record>
+</odoo>
